Cleaned up and documented StringCompatibility
